build-runtime.py: Add support for overlay suites
This lets us build Steam Runtime variants with proposed or test packages
from a third-party apt repository, without having to copy the entire
official Steam Runtime package set into that repository (and update it
when the official Steam Runtime repository is updated).

The [trusted=yes] option is currently parsed but ignored: we download
files from the apt repository directly, and we do not implement
secure-APT verification for them. Only use this script if you have a
secure path to the apt repository, for example https, local access or
a VPN.

Signed-off-by: Simon McVittie <smcv@collabora.com>