build-runtime: Only install debug symbols if they match the library
pkg-create-dbgsym can be rather picky about the build environment. If we
don't generated new detached debug symbols for a package for some reason,
we can end up with outdated debug symbols lingering in the apt
repository; but there is no point in installing those, because they
won't help to debug the mismatched shared library.

Signed-off-by: Simon McVittie <smcv@collabora.com>