build-runtime.py: Require that runtime doesn't exist and templates does
If files in ./templates (or the argument to --templates if different)
are changed or added, while reusing an existing ./runtime (or the
argument to --runtime if different), then we would unexpectedly fail to
update those files. For example, after fixing a bug in run.sh, we
wouldn't pick up the fixed version in the runtime until we either use
a different target directory or delete and recreate ./runtime.

Similarly, if files in ./templates are removed, their counterparts in
./runtime would stay until we either switch to a different target
directory or delete ./runtime.

If we require the caller to give us a nonexistent target directory, it
becomes completely clear what is going on, and makes the caller
responsible for doing the right thing (either using a different target
directory for each build, deleting unwanted old builds, or moving
unwanted old builds out of the way).

While we're here, also validate that ./templates exists: if it didn't,
our call to shutil.copytree() would just have crashed anyway, but we
might as well have a nicer error message in the common case.

Signed-off-by: Simon McVittie <smcv@collabora.com>