packages: Include libmpg123 shared libraries, not mpg123 executable
It's important that we distinguish between source packages (the first
column of this file) and binary packages (every other column).

libsdl2-mixer dlopens libmpg123.so.0 to play MP3 audio, so we want that
in the Runtime, otherwise games that use SDL for MP3 audio won't work
unless the host system happens to have libmpg123-0 (or equivalent)
installed on it.

We don't want /usr/bin/mpg123, the command-line MP3 player, which is
useless without libmpg123.

Similarly, we presumably want development runtimes to have libmpg123-dev,
the development headers for this MP3 playback library, rather than the
nonexistent mpg123-dev.

I've assumed that we don't need libout123, which appears to be an audio
output abstraction layer analogous to SDL2 or PortAudio, used by the
mpg123 command-line MP3 player.

Signed-off-by: Simon McVittie <smcv@collabora.com>