sourcepkgs: Update from live infrastructure
The libsdl2-mixer in scout_beta requires libraries from mpg123, opus
and opusfile. These were listed in packages.txt but not in
sourcepkgs.list, causing `make check` to fail.

sourcepkgs.list is the list of source packages to be included in the
apt repository for the runtime, and this change has already been deployed
on the live Steam Runtime infrastructure
<http://repo.steampowered.com/steamrt/conf/sourcepkgs.list>.

Signed-off-by: Simon McVittie <smcv@collabora.com>