build-runtime: Rename --runtime to --output, and take away the default
Prior to commit 31764950, the --runtime directory (default: ./runtime)
was both an input and an output, and was always populated.

Now that it's only an output, with --templates=./templates taking the
input role, rename the option to --output to make this more clear.

While I'm making incompatible changes anyway, instead of giving it a
new default, I've given it no default. Instead, we require at least one
of --output and --archive. If the only thing we are interested in
generating is a tar archive (as is the case for the official Steam
Runtime binary releases provided alongside Steam), then we can build
the runtime in a temporary directory, compress it into the archive, and
discard the temporary directory, removing the need for a wrapper script
to manage the temporary directory.

Signed-off-by: Simon McVittie <smcv@collabora.com>