build-runtime: Improve normalization for paths in archive
When sorting archive members by name, we normalized `foo` to `./foo`
as intended, but also normalized `.` to `./.` not as intended. This
resulted in a predictable but odd archive order where the immediate
children of the top-level directory (including subdirectories like
amd64/ and scripts/) were all sorted before the contents of any
subdirectory (including for example amd64/).

Signed-off-by: Simon McVittie <smcv@collabora.com>