Remove sourcepkgs.list
This was required when only a subset of the packages built by the OBS
autobuilder were copied into the production apt repository. We had
to be careful that the source package list was consistent with the
binary package list, to avoid causing bugs like the one fixed in #132.

However, the new plan is that this filtering will no longer be done, and
all built packages will be mirrored into the production apt repository
(this will result in a few new packages appearing, such as automake1.15).

Signed-off-by: Simon McVittie <smcv@collabora.com>