setup_chroot.sh: Add support for importing prebuilt sysroot tarballs
Making the debootstrap process safe to run on arbitrary user systems
is relatively involved, and we can get more predictable results by
distributing a complete tarball that corresponds 1:1 to a specific
version of a runtime. If we build Docker images, the sysroot tarball
is essentially the same as the contents of the Docker image.

Signed-off-by: Simon McVittie <smcv@collabora.com>