build-runtime.py: Include source packages without unpacking and repacking
This has a somewhat worse compression ratio because xz can't take
advantage of redundancy between similar source code files, but is much
faster than unpacking and repacking. With xz compression, a Runtime with
source code was previously 458M and took 12:49; on the same hardware it
is now 558M and takes 5:54.

With gz compression, it is both faster (2:19 -> 0:50) and smaller
(654M -> 601M) to include the source packages this way.

Signed-off-by: Simon McVittie <smcv@collabora.com>