templates/setup.sh: Improve robustness of quoting
The shellcheck in Debian 10 'buster' doesn't warn about this, but the
version in Debian 9 'stretch' does. The code does the right thing
either way (it expands to the equivalent of either `[ ! -z ]`, which
fails as desired, or `[ ! -z isset ]`, which succeeds as desired), but
it's more obviously correct if we quote it, turning the first case into
`[ ! -z "" ]` (which also fails, as desired).

Signed-off-by: Simon McVittie <smcv@collabora.com>