build-runtime.py: Add support for building from metapackages
Instead of duplicating the lists of desired packages in the steamrt
source package and in packages.txt, it would be useful to be able to
have a "single source of truth". When testing new packages we need an
extra apt source (--extra-apt-source) to fetch those packages from
anyway, so it seems natural to add an updated steamrt source package to
that repository and use it as our single source of truth.

This will mean we can add or update new batches of packages, like the
GNUTLS 3 transition in #101, without temporarily breaking the public
version of steam-runtime.git for third-party users by making it rely
on packages that have not yet gone out to the public.

With the modified steamrt package that I'm preparing to accompany this
change, the metapackage equivalent of the default behaviour for scout
(packages.txt) becomes:

    --metapackage steamrt-libs
    --metapackage steamrt-legacy

or for --debug mode:

    --debug
    --metapackage steamrt-libs
    --metapackage steamrt-legacy
    --metapackage steamrt-libdevel
    --metapackage steamrt-libdevel-non-multiarch
    --metapackage dummygl-dev

Similarly, the equivalent of --packages-from=webhelper/packages.txt
in its Debian-8-derived suite is just this:

    --metapackage steamrt-libs

To check that this change doesn't cause regressions by altering the
contents of the runtime tarball, it is now possible to specify both
the --metapackage and --packages-from options. The script will assume
that they are meant to be exactly equivalent, and will issue warnings
for any situations where using one but not the other would have made
a difference.

This is not a full, recursive dpkg dependency resolver: we only include
the metapackages themselves, their direct dependencies, and a minimal
number of three-level dependency chains selected by
accept_transitive_dependency() (notably
steamrt-libs -> libgcc1 -> gcc-x.y-base). However, the script will
issue warnings, which can be made fatal with --strict, if there are
additional second- or third-level dependency chains that are not
explicitly ignored by ignore_metapackage_dependency() or
ignore_transitive_dependency().

Signed-off-by: Simon McVittie <smcv@collabora.com>