Force libcurl.so.[34] to be taken from the Steam Runtime
libcurl in the Steam Runtime has a Debian-specific oddity (inherited by
Ubuntu and SteamRT) in which libcurl.so.3 is a symlink to libcurl.so.4,
as a result of some unfortunate ABI weirdness back in 2007 (upstream
bumped the SONAME in a situation where it might not in fact have been
necessary).

It also has Debian-specific symbol versioning, as a result of the
versioned symbols introduced as a Debian-specific change in 2005-2006
being preserved across the rename from libcurl.so.3 to libcurl.so.4
in 2007, and not matching the versioned symbols that upstream
subsequently added (in 2011). Versions of Debian and Ubuntu between
2011 and 2018 (including Ubuntu 12.04, the basis for the Steam Runtime)
had a patched libcurl to remain compatible with their 2007 version.

In 2018, Debian and Ubuntu switched to a SONAME and versioning that match
upstream, but the Steam Runtime is based on a version that is older than
that, so anything built against the Steam Runtime will expect the
2007-Debian-compatible versioned symbols. Make sure we use the Steam
Runtime version, which will satisfy those games' expectations.

In particular, this is known to affect Risk of Rain, and perhaps other
GameMaker-based games.

Signed-off-by: Simon McVittie <smcv@collabora.com>