build-runtime: Also look for debug symbols in normal Packages file
In Ubuntu, automatic -dbgsym packages are generated with the .ddeb
extension.

The apt repositories currently available on
<http://repo.steampowered.com/> were built with a modified reprepro
originating from <https://bugs.debian.org/730572>, which diverts
ddebs to a pseudo-component like "main/debug", similar to the existing
handling of udebs. Most of the changes from #730572 were never accepted
into upstream reprepro.

To avoid maintaining a patched reprepro indefinitely, it might be a good
idea to change debhelper so it puts the -dbgsym packages in ordinary
.deb files. This change makes build-runtime look in both package lists
so that it can cope either way.

Signed-off-by: Simon McVittie <smcv@collabora.com>