build-runtime: Skip more Mesa libraries
Some diagnostic tools that I'm looking at adding to the Steam Runtime
would pull in EGL and GLES libraries in addition to GLX and
"desktop GL". We want all of these to be taken from the host system.

While I'm here, also add the various Mesa ICDs used with GLVND-style
loaders. We don't want to include those in the Runtime either.

Signed-off-by: Simon McVittie <smcv@collabora.com>