build-runtime: Combine all directories that we can
steam-runtime-tools expects to find the ABI expectations and the
"helper" tools that it uses in a single directory each, so it benefits
from at least usr/lib/steamrt and usr/libexec/steam-runtime-tools-0
being combined.

However, recent Steam Runtime development has been careful to only
include files that are multiarch-compatible and do not collide, so
there isn't really any reason not to merge the whole tree, except for
usr/bin (and in theory bin, sbin, usr/sbin) which contains executables
for each architecture.

This saves approximately 25% of the compressed size of the Steam
Runtime (when using gzip).

Signed-off-by: Simon McVittie <smcv@collabora.com>