run.sh: Drop redundant directories from LD_LIBRARY_PATH
Now that amd64/lib and i386/lib are the same directory, and likewise
for amd64/usr/lib and i386/usr/lib, we can simplify the LD_LIBRARY_PATH.

Move /lib and /usr/lib to the end of the search path. This is
consistent with the search order on a multiarch Debian system, where
all of the multiarch library directories take precedence over the
unspecified-architecture directories.

Signed-off-by: Simon McVittie <smcv@collabora.com>