Ignore i386 packages in the amd64 Packages file and vice versa
The apt repositories generated for personal branches in the Open Build
Service have separate i386 and amd64 Packages files, but each lists
packages for both architectures. When installing libraries for i386,
we need to skip the amd64 stanzas and vice versa.

As a special case, architecture-independent data files (Architecture: all)
continue to be considered to be a valid part of every architecture.

Signed-off-by: Simon McVittie <smcv@collabora.com>