setup.sh: Use command -v instead of which
The command -v shell builtin is mandated by POSIX, so it's more
predictable and more likely to exist than which, which is non-standard:
a version from the debianutils package is considered Essential on
Debian-derived systems, a GNU implementation is also popular, and some
shells have it as a builtin, but it isn't necessarily ubiquitous.

command -v doesn't produce output on failure, avoiding the need for a
redirection.

Signed-off-by: Simon McVittie <smcv@collabora.com>
Shellcheck: SC2230