Use https for all URLs, except those that are secured by apt
As a general principle, we should have a trusted path to everything we
download, even if the network is malicious.

We don't need this for debos-chroot.yaml, scripts/bootstrap-runtime.sh,
setup_chroot.sh because they authenticate the packages they download
by using PGP keys (see apt-secure(8)). In older Debian derivatives
apt can't handle https on its own, and using https makes it harder to
go via a caching proxy, so it's undesirable to use https just on
principle.

Signed-off-by: Simon McVittie <smcv@collabora.com>