run.sh: Include the external LD_LIBRARY_PATH before the runtime libraries
If a user wants to launch Steam, or a game, with a custom
LD_LIBRARY_PATH, we should search these paths right after the pinned
libraries. Otherwise the runtime libraries will have an higher priority
than the libs provided by the users, and this is likely not the expected
behavior.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>