Update bug reporting information
Now that people are encountering pressure-vessel/SteamLinuxRuntime via
Proton 5.13, we're getting more bug reports from soldier than scout,
but this document was only really valid for scout.

Because of the way container setup is currently done for Proton games,
the method involving Launch Options documented in #270 is less useful
than it used to be: the launch options aren't processed until we get
as far as launching the actual game, which is too late.

Signed-off-by: Simon McVittie <smcv@collabora.com>