setup: Speed up libraries pinning
One of the slowest operation during the setup is calling `file -L` to
check if a particular library is 32-bit or 64-bit.

To speed up the process we reduce the number of external calls to `file
-L` by storing the result and avoiding this check entirely when not
necessary.

On my system the library pinning was taking about 11.27 seconds, and
with this optimization it went down to 7.55 seconds.
Because I did my testings with a recent desktop CPU and an NVME, it's
safe to assume that the gain will be even more noticeable on less
powerful systems.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>