setup: Use parameter expansion instead of basename
By switching to the bash parameter expansion we have a speedup of nearly
1 second on recent desktop systems with NVME, and probably more on
slower systems.

They are not exactly the same but here we are using a somewhat
controlled input, the result of `find` for the libraries in the absolute
path of the steam runtime, so it should not give us unexpected results.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>