known-issues: Generalize info about games that reset LD_LIBRARY_PATH
This is known to affect the Hashlink interpreter, but is not unique to
the Hashlink interpreter: Game Dev Tycoon (#198) has a similar issue.

To give developers a hint, I've given an example of a startup script
that won't work in the container runtime, and a version of it that will.

Signed-off-by: Simon McVittie <smcv@collabora.com>