run.sh: Handle STEAM_COMPAT_FLAGS environment variable
Currently the only expected flag in `STEAM_COMPAT_FLAGS` is
`search-cwd`, that is used to append the game current working directory,
stored in `STEAM_COMPAT_INSTALL_PATH`, to the `LD_LIBRARY_PATH`.

Previously we handled this flag in pressure-vessel, but this may give
`STEAM_COMPAT_INSTALL_PATH` an higher priority than expected.
For example, when we are using the new layered scout runtime,
`STEAM_COMPAT_INSTALL_PATH` will be searched before the container
runtime's linker paths and the Steam Runtime's library paths.

Co-authored-by: Simon McVittie <smcv@collabora.com>
Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>