templates: Add a script to bootstrap runtimes
This will initially be used for heavy, but it's general enough that it
could in principle be used for scout, replacing parts of steam.sh.

When `--unpack-dir` is not given, the same script can also be used to
"escape" from the Steam Runtime environment, by specifying `--runtime=""`,
or to encapsulate the setup and run phases of a custom runtime during
testing, by specifying `--runtime`.

Signed-off-by: Simon McVittie <smcv@collabora.com>