templates: Consistently use stderr for diagnostic messages and progress
Some app-launchers run Steam with stdout pointing to the write end of a
pipe to something that will process it in some machine-readable way, or
the write end of a pipe that has nothing at the read end at all. This
means that when we write to stdout, our output can be misinterpreted as
something machine-readable in the former case, or fail with "Broken pipe"
(`EPIPE`) in the latter case.

This is arguably a bug in those app-launchers or the way they are
configured, because there is no guarantee that arbitrary applications
won't write to stdout in this way; but we can mitigate it by consistently
using stderr for human-readable diagnostic messages and human-readable
progress indicators.

The zenity_progress code path in `setup.sh` `pin_newer_runtime_libs`
continues to use stdout, because in the zenity_progress code path, it
is a machine-readable progress indication that is piped to zenity.

`run.sh --print-steam-runtime-library-paths` continues to use stdout,
because it is providing a machine-readable list of library paths for
its caller.

Similarly, `setup.sh --print-bin-path` continues to use stdout, because
it is providing a machine-readable list of executable paths for its
caller.

This is likely to resolve
https://github.com/ValveSoftware/steam-for-linux/issues/8069 and
related issues.

Signed-off-by: Simon McVittie <smcv@collabora.com>