setup.sh: Only consider host libraries from highest-precedence directory
On some host operating systems, notably Gentoo, each version of gcc
installs its own copy of libstdc++.so.6 into a version-specific directory
such as /usr/lib/gcc/x86_64-pc-linux-gnu/11.1.0, with those directories
added to the ldconfig search path in descending version order. ld.so
will use the first one it sees, but until now, setup.sh used the *last*
one it saw when deciding whether to pin the Steam Runtime libstdc++.so.6.

If an old version like gcc 4.9 is still installed, then we will think
the Steam Runtime's libstdc++.so.6 from gcc 5 is newer than the host's
libstdc++.so.6 from gcc 4.9, causing the version from gcc 5 to be pinned;
but that version will not have all the symbols required for the host's
graphics drivers, causing a runtime link failure.

Instead, we should compare the *first* host libstdc++.so.6 with the
runtime's libstdc++.so.6, which will give us the right answer: do not
pin the runtime's libstdc++.so.6, unless the host OS is something
particularly old (Debian 8, Ubuntu 14.04, SteamOS 2).

This should resolve
<https://github.com/ValveSoftware/steam-for-linux/issues/8123>.

Signed-off-by: Simon McVittie <smcv@collabora.com>