possible-designs: Refer to layered LDLP runtime as used for heavy
For a while, the steamwebhelper was using a simpler LD_LIBRARY_PATH
arrangement, but we eventually switched it to this design to avoid
compatibility issues with newer host OSs.

We could in theory still do this for games, but we have chosen not to:
the heavy runtime is incomplete (only enough to run the steamwebhelper),
and a more modern LD_LIBRARY_PATH runtime would break compatibility with
older host OSs and require more maintenance effort (patching) than is
really sustainable. A more likely route to get native Linux games
running on a newer stack would be to enable use of the soldier or sniper
container runtimes, which Proton already relies on.

Signed-off-by: Simon McVittie <smcv@collabora.com>