README: Point to Steam Runtime SDK for details of how to build games
Because of the way Gitlab's integrated Docker registry works, we need
to have a project for each branch of the Steam Runtime SDK, so that's a
natural place to put reference documentation for how to use it.
Outsource the developer documentation to there, so that this repository
can have more focus on debugging end-user-facing runtime issues.

Docker, Podman and Toolbx share many advantages, but there are some
tradeoffs between them, so for now I'm recommending Docker and Podman
equally:

- Docker is available on older host operating systems, and is
  sufficiently ubiquitous that it needs little introduction in 2022

- Podman doesn't require root privileges and interoperates with Toolbx,
  but is less widespread in LTS distributions: at the time of writing,
  Debian stable has it, but Ubuntu LTS does not

- Toolbx is not currently supported by the GA version of scout, and
  is in fewer LTS distributions than Podman, but provides a lot more
  developer convenience than Docker or Podman, more like the UX of
  the old schroot environment

When the current beta of Steam gets promoted to GA, bringing a new
scout release with it, we'll be able to recommend Toolbx more
wholeheartedly.

Signed-off-by: Simon McVittie <smcv@collabora.com>