README: Add a more user-facing introduction
Talking about chroots is useful for game developers, but we should
introduce what the runtime is before we get into that level of detail.

What I'm intending to do with this is to have a relatively brief
user-facing introduction here, and then outsource detailed technical
information to steam-runtime-tools, where it can be updated at the same
time as the matching code changes.

Signed-off-by: Simon McVittie <smcv@collabora.com>