doc: Update to reflect SteamOS 3 not being Debian-derived
SteamOS versions 1 and 2 were derived from older releases of Debian, so
we have generally grouped it together with older Debian releases when
talking about compatibility. It is now public that SteamOS 3, as used on
the Steam Deck, is derived from Arch Linux instead, and as a result it
behaves more like a recent version of Arch Linux than like an old version
of Debian; update documentation to match this.

Signed-off-by: Simon McVittie <smcv@collabora.com>