run.sh: Use pressure-vessel's ld.so.conf, if found
In some distributions, like Solus, ldconfig defaults to reading a path
that is not the interoperable /etc/ld.so.conf.

pressure-vessel relies on having ldconfig read the ld.so.conf that it
has written out, so to be fully robust, ideally it would know about all
these distro-specific paths and replace them all with symlinks to the
correct content, as it already does for /etc/ld.so.conf. However, we
can force the issue by detecting pressure-vessel and applying some
"do what I mean".

Signed-off-by: Simon McVittie <smcv@collabora.com>