setup.sh: Be more resilient against interruption
If setup.sh is interrupted or fails, the conservative thing to do is to
redo pinning the next time we start up. However, until now, it would
interpret empty pinned_libs_ directories as indicating that no
libraries needed to be pinned and nothing needed re-checking.

To avoid that, create a flag file `pinned_libs_*/done` when we are
finished dealing with each word size, and always redo pinning if that
flag file does not exist.

Signed-off-by: Simon McVittie <smcv@collabora.com>