setup.sh: Optionally use libcurl-compat-setup to provide two libcurl ABIs
If glibc and steam-runtime-tools are both new enough versions, we can
set up a shim version of libcurl that provides both the historical
Debian ABI (which is scout-compatible) and the upstream ABI, allowing
tools like Arch's build of valgrind (with debuginfod support using
libcurl) to run inside the Steam Runtime environment.

See steamrt/steam-runtime-tools!491 for more details of the ABI history
that led to this, and the helper tool and shim libcurl library that it
requires.

Because this involves some fairly obscure details of how glibc implements
ELF and versioned symbols, it's currently gated behind an environment
variable check to avoid perturbing the ABI seen by ordinary games. The
only safe assumption for games that run in the LD_LIBRARY_PATH runtime is
that every library is compatible with, and at least the same version as,
the versions in scout; anything beyond that is quality-of-implementation.

Helps: steamrt/tasks#119
Signed-off-by: Simon McVittie <smcv@collabora.com>