run.sh: Use libcurl shim if running Steam under valgrind
Recent versions of valgrind are linked to libdebuginfod, and therefore
indirectly libcurl, so we need to be able to use the host libcurl even if
it's incompatible with the scout libcurl. steam.sh uses the environment
variable DEBUGGER as an instruction to run Steam itself under valgrind
or another debugging tool.

Signed-off-by: Simon McVittie <smcv@collabora.com>