setup.sh: Also pin libcurl-gnutls.so.4 to our version
libcurl-gnutls.so.4 has the same compatibility issues as libcurl.so.4 -
if anything more so, because the Debian/Ubuntu ABI of libcurl.so.4
re-converged with upstream in 2018, but libcurl-gnutls.so.4 is still
versioned CURL_GNUTLS_3 in Debian/Ubuntu and CURL_GNUTLS_4 upstream
as of mid 2022.

This was mitigated by the fact that most OSs either don't ship a GNUTLS
build of libcurl at all (e.g. Red Hat), or ship it without versioned
symbols (e.g. Arch), or ship a Debian-compatible version because they
are Debian-derived themselves. However, OpenMandriva appears to provide
a GNUTLS build of libcurl that is upstream-compatible rather than
Debian-compatible.

Resolves: https://github.com/ValveSoftware/steam-runtime/issues/535
Signed-off-by: Simon McVittie <smcv@collabora.com>