setup.sh: Always pin some libraries even if not found on host system
Previously, we would only put symlinks in the `pinned_libs_*` directories
if we found a corresponding SONAME on the host system.  However,
the continuing ABI uncertainty around libcurl makes this unreliable:
we can find that there is a library on the host system that provides
one of the filenames that we expect to be a SONAME, but internally has
some different string as its SONAME. In particular, on OpenMandriva,
some versions of the library that is available as libcurl-gnutls.so.4
actually had libcurl.so.4 as their ELF DT_SONAME header, which we wouldn't
be able to find in ld.so.cache under the name libcurl-gnutls.so.4.

Simplify this by symlinking forced libraries into `pinned_libs_*`,
regardless of whether they appear to be present on the host system
or not. Either way, they'll be found via LD_LIBRARY_PATH search.

Resolves: https://github.com/ValveSoftware/steam-runtime/issues/535
Signed-off-by: Simon McVittie <smcv@collabora.com>