setup.sh: Don't pin libdbus-1.so.3 or libgtk-x11-2.0.so.0 by default
These libraries are not tightly coupled to libdbusmenu-glib.so.4 or to
libdbusmenu-gtk.so.4, which are already expected to cope with different
versions of the lower-level libdbus-1.so.3 and libgtk-x11-2.0.so.0.

libdbus-1.so.3 is still maintained, and sometimes adds new API, which
would not have been available to 32-bit processes running under the
Steam Runtime before this change. In particular, this could become a
problem for host-system copies of SDL or Pipewire which might require
newer libdbus APIs.

libgtk-x11-2.0.so.0 is end-of-life, but the version found on typical
host systems will be newer than the one in scout. Pinning it also causes
a potential problem with mismatches between libgtk-x11-2.0.so.0 and
libgdk-x11-2.0.so.0, which are from the same source package and therefore
can have private APIs between them, similar to libdbusmenu-glib.so.4
and libdbusmenu-gtk.so.4.

Just in case this causes regressions on some unusual system, provide
environment variables which can be used to force these libraries to be
pinned as they historically were. These will be removed later if they
do not seem to be necessary in practice.

Like `STEAM_RUNTIME_USE_LIBCURL_SHIM`, the environment variables take
effect in run.sh, not in setup.sh; this allows them to take effect
without having to repeat setup.

Helps: steamrt/tasks#256
Signed-off-by: Simon McVittie <smcv@collabora.com>