setup, run: Always pin 32-bit libgtk-x11-2.0.so.0 if a system copy exists
It seems that our 32-bit version of libgtk-x11-2.0.so.0 is also
necessary for Steam's tray icon to work as intended, so we will have to
live with sticking to a 2012 version of GTK 2 for a while longer.

This partially reverts commit 09dec78c "setup.sh: Don't pin libdbus-1.so.3
or libgtk-x11-2.0.so.0 by default".

Signed-off-by: Simon McVittie <smcv@collabora.com>