setup.sh: Don't fail if stdout is a broken pipe
Similar to the issue fixed in commit 2cc8124e "scripts: Ignore errors
when showing diagnostic messages", if zenity crashes or is broken,
writing to its standard input will fail. We're running under `set -e`,
so that will cause the whole script to abort unless we suppress that
behaviour.

The progress reporting is only a "nice to have", so don't let it stand
in the way of this script running at all.

steamrt/tasks#289, mitigates:
https://github.com/ValveSoftware/steam-for-linux/issues/9356

Signed-off-by: Simon McVittie <smcv@collabora.com>