setup.sh: Treat libxcrypt libcrypt.so.1 as newer than host glibc version
libxcrypt produces a drop-in replacement for glibc `libcrypt.so.1`
when compiled with `--enable-obsolete-api=glibc`, as it is in for
example Debian.

libxcrypt's `libcrypt.so.1 -> libcrypt.so.1.1.0` is not easily comparable
with glibc's `libcrypt.so.1 -> libcrypt-2.28.so` (or similar); but we
know that by design, libxcrypt's `libcrypt.so.1` exports every symbol
that was ever included in glibc's `libcrypt.so.1`, so we can treat it
as though it was always strictly newer.

This will allow us to ship a `libcrypt.so.1` in the Steam Runtime
LD_LIBRARY_PATH runtime, and rely on it being used on any host system
that either does not install `libcrypt.so.1` or is still using glibc's
`libcrypt.so.1`, while continuing to use the host `libcrypt.so.1` if the
host copy is actually from libxcrypt.

steamrt/tasks#332

Signed-off-by: Simon McVittie <smcv@collabora.com>