build-runtime: Treat architectures as ordered, with first = primary
We now have some packages that overlap between architectures, notably
steam-runtime-tools-minimal which includes
usr/libexec/steam-runtime-tools-0/srt-bwrap and srt-logger.
At the moment, we keep the version of these from whichever architecture
happens to be unpacked first, which is the one that comes first in
alphabetical order.

We can make this more predictable by explicitly choosing a primary
architecture. Now that Steam requires 64-bit support anyway (for
steamwebhelper), there's no problem with treating amd64 as the primary
architecture that we must be able to run whatever else happens.

This actually has no practical effect on the contents of our runtimes,
because amd64 happens to sort before i386 in alphabetical order, but
it makes it clearer that what we're doing is intentional!

Signed-off-by: Simon McVittie <smcv@collabora.com>