build-runtime: Don't warn for files known to overlap between architectures
For historical reasons we split up usr/bin into amd64 and i386 versions,
but we do not do this for usr/libexec, which results in some executables
only existing for amd64, with their i386 counterpart discarded.

In situations where we know this is happening and know that it's
benign, there's no need to warn about this.

Signed-off-by: Simon McVittie <smcv@collabora.com>