build-runtime: Create extra symlinks in game-bin/
These executables all launch URLs or commands outside the Steam Runtime
environment using an IPC mechanism (D-Bus messages to xdg-desktop-portal,
or messages through steam.pipe to Steam). They can be used to implement
operations like "open a URL" in Steam games, without inheriting aspects
of the game's execution environment that are likely to be undesirable,
such as the Steam Runtime LD_LIBRARY_PATH, the game overlay LD_PRELOAD,
the game subreaper, non-default rlimits, and so on.

(For example, if opening a URL results in Firefox being a child process
of Portal, then Steam will not think that Portal has exited until
Firefox exits, and clicking on the Stop button in Steam will terminate
Firefox, which is unlikely to be what anyone wanted.)

To avoid infinite recursion (Steam opening a URL via one of these tools,
which tries to open the same URL via Steam) this directory must not be
added to the PATH for Steam itself, but only for individual games.

steamrt/tasks#489

Signed-off-by: Simon McVittie <smcv@collabora.com>