README: Link to Steam Runtime SDK docs instead of repeating ourselves
Adding and updating detailed documentation in this repository is
inconvenient, because for historical reasons its git history contains
large binaries, which cannot be removed without invalidating branches
and forks of the repository.

Each branch of the Steam Runtime has its own SDK hosted elsewhere,
so link to those as the primary thing to use.

Signed-off-by: Simon McVittie <smcv@collabora.com>